refactor(frontend): extract shared voyage components out of VoyageDetailPage
Move the reusable pieces of the 1823-line VoyageDetailPage into
features/voyages/shared/ so the upcoming experiment console can use them
without a third copy-paste:

- stepUtils.ts: byListOrder, stepMarker, stepTokenCount, plan-history
  helpers and the timeline entry builder
- StepCard.tsx: StepCard with its acceptance/attempts/observation blocks,
  the plan-adjustment separator card, and the wiki/experiment
  human-readable step summaries
- ActivityBar.tsx: activity one-liner, mode badge, paused-state banners
- terminal.tsx: the dark TaskTerminal family (LogLine/LlmRecord/LlmActive),
  now with optional height/extraEntries/footer/headerExtras/filter props
  (defaults keep the previous behavior bit-for-bit)
- useVoyageChannel.ts: SSE subscription + throttled terminal buffer +
  history replay + status/step cache merge, with an onExtraEvent
  passthrough for future message/ask events

VoyageDetailPage keeps only its page-specific parts (presentation
download, wiki run summary, layout) and imports the rest. Pure move,
zero behavior change; tsc and vite build pass.

Closes #316
